refactor(runner/scheduler): split mode-specific paths
