+++ packages/runner/src/scheduler/events.ts
@@ -185,7 +185,6 @@ export function addSchedulerEventHandler(state: {
 export interface SchedulerEventExecutionState {
   readonly runtime: Runtime;
   readonly eventQueue: QueuedEvent[];
-  readonly pullMode: boolean;
   readonly dirty: ReadonlySet<Action>;
   readonly pending: Set<Action>;
   readonly eventPreflightTelemetryEnabled: boolean;
@@ -214,95 +213,6 @@ export interface SchedulerEventExecutionState {
   readonly snapshotDirtyDependencyTraceContext: (
     trace: DirtyDependencyTraceContext,
   ) => SchedulerEventPreflightStats;
-}
-
-export async function processQueuedEventDuringExecute(
-  state: SchedulerEventExecutionState,
-  eventBlockingDeps: Set<Action>,
-): Promise<void> {
-  // Process next event from the event queue.
-  const queuedEvent = state.eventQueue[0];
-  if (!queuedEvent) return;
-
-  if (
-    queuedEvent.notBefore !== undefined &&
-    queuedEvent.notBefore > performance.now()
-  ) {
-    state.scheduleEventQueueWake(queuedEvent.notBefore);
-    return;
-  }
-
-  delete queuedEvent.notBefore;
-
-  const { handler } = queuedEvent;
-  const handlerId = state.getActionId(handler);
-
-  // In pull mode, ensure handler dependencies are computed before running.
-  let shouldSkipEvent = false;
-  if (state.pullMode && handler.populateDependencies) {
-    const preflight = preflightQueuedEventDependencies({
-      runtime: state.runtime,
-      eventQueue: state.eventQueue,
-      dirty: state.dirty,
-      pending: state.pending,
-      pendingActions: state.pending,
-      eventBlockingDeps,
-      handleError: (error, target) => state.handleError(error, target),
-      setDirtyDependencyTraceContext: (trace) => {
-        state.setDirtyDependencyTraceContext(trace);
-      },
-      collectDirtyDependenciesForLog: (deps, dirtyDeps, dirtyDepMemo) =>
-        state.collectDirtyDependenciesForLog(
-          deps,
-          dirtyDeps,
-          dirtyDepMemo,
-        ),
-      isDebouncedComputationWaiting: (dep) =>
-        state.isDebouncedComputationWaiting(dep),
-      getNextDebounceRunTime: (dep) => state.getNextDebounceRunTime(dep),
-      getNextEligibleRunTime: (dep) => state.getNextEligibleRunTime(dep),
-      scheduleEventQueueWake: (notBefore) =>
-        state.scheduleEventQueueWake(notBefore),
-    }, queuedEvent);
-    shouldSkipEvent = preflight.shouldSkipEvent;
-
-    if (state.eventPreflightTelemetryEnabled) {
-      state.runtime.telemetry.submit({
-        type: "scheduler.event.preflight",
-        handlerId,
-        handlerInfo: state.getActionTelemetryInfo(handler),
-        readCount: preflight.deps.reads.length,
-        shallowReadCount: preflight.deps.shallowReads.length,
-        dirtySizeBefore: preflight.dirtySizeBefore,
-        pendingSizeBefore: preflight.pendingSizeBefore,
-        dirtyDependencyCount: preflight.dirtyDeps.size,
-        hasDirtyDependencies: preflight.hasDirtyDependencies,
-        skipped: shouldSkipEvent,
-        populateMs: preflight.populateMs,
-        txToLogMs: preflight.txToLogMs,
-        depCommitMs: preflight.depCommitMs,
-        collectMs: preflight.collectMs,
-        scheduleMs: preflight.scheduleMs,
-        stats: state.snapshotDirtyDependencyTraceContext(
-          preflight.preflightStats,
-        ),
-      });
-    }
-  }
-
-  if (shouldSkipEvent) return;
-
-  await dispatchQueuedEvent({
-    runtime: state.runtime,
-    eventQueue: state.eventQueue,
-    setRunningPromise: (promise) => {
-      state.setRunningPromise(promise);
-    },
-    getActionId: (target) => state.getActionId(target),
-    getActionTelemetryInfo: (target) => state.getActionTelemetryInfo(target),
-    handleError: (error, target) => state.handleError(error, target),
-    queueExecution: () => state.queueExecution(),
-  }, queuedEvent);
 }
 
 export function preflightQueuedEventDependencies(state: {
