+++ packages/runner/src/scheduler/notifications.ts
@@ -2,9 +2,7 @@ import type { MemorySpace } from "@commonfabric/memory/interface";
 import type {
   ChangeGroup,
   IMemoryChange,
-  IStorageSubscription,
   IStorageTransaction,
-  StorageNotification,
 } from "../storage/interface.ts";
 import type { TriggerIndexState } from "./trigger-index.ts";
 import { summarizeTriggerTraceValue } from "./diagnostics.ts";
@@ -16,11 +14,7 @@ import type {
   TriggerTraceScheduledEffect,
 } from "./types.ts";
 
-type SchedulerMode = "pull" | "push";
-
-export function schedulerMode(pullMode: boolean): SchedulerMode {
-  return pullMode ? "pull" : "push";
-}
+export type SchedulerMode = "pull" | "push";
 
 export function hasRegisteredTriggers(
   state: TriggerIndexState,
@@ -44,7 +38,7 @@ export function createTriggerTraceEntry(state: {
   readonly notificationType: string;
   readonly changeIndex: number;
   readonly matchedActionCount: number;
-  readonly pullMode: boolean;
+  readonly mode: SchedulerMode;
   readonly writerActionId?: string;
   readonly space: MemorySpace;
   readonly change: IMemoryChange;
@@ -55,7 +49,7 @@ export function createTriggerTraceEntry(state: {
     notificationType: state.notificationType,
     changeIndex: state.changeIndex,
     matchedActionCount: state.matchedActionCount,
-    mode: schedulerMode(state.pullMode),
+    mode: state.mode,
     writerActionId: state.writerActionId,
     space: state.space,
     entityId: state.change.address.id,
@@ -125,7 +119,7 @@ export function planPullTriggeredAction(
 export function createTriggerTraceActionRecord(state: {
   readonly actionId: string;
   readonly actionType: "effect" | "computation";
-  readonly pullMode: boolean;
+  readonly mode: SchedulerMode;
   readonly decision: TriggerTraceActionRecord["decision"];
   readonly pendingBefore: boolean;
   readonly pendingAfter: boolean;
@@ -136,7 +130,7 @@ export function createTriggerTraceActionRecord(state: {
   return {
     actionId: state.actionId,
     actionType: state.actionType,
-    mode: schedulerMode(state.pullMode),
+    mode: state.mode,
     decision: state.decision,
     pendingBefore: state.pendingBefore,
     pendingAfter: state.pendingAfter,
@@ -152,23 +146,7 @@ export function shouldRecordTriggerTraceEntry(
   return entry.triggered.length > 0 || entry.matchedActionCount > 0;
 }
 
-function planTriggeredNotificationAction(state: {
-  readonly pullMode: boolean;
-  readonly isEffect: boolean;
-  readonly dirtyBefore: boolean;
-  readonly isOwnCommitSource: boolean;
-  readonly hasSourceChangeGroup: boolean;
-  readonly actionChangeGroup: ChangeGroup | undefined;
-  readonly sourceChangeGroup: ChangeGroup | undefined;
-}): TriggeredActionPlan {
-  if (state.pullMode) {
-    return planPullTriggeredAction(state);
-  }
-
-  return planPushTriggeredAction(state);
-}
-
-function applyPushTriggeredActionPlan(
+export function applyPushTriggeredActionPlan(
   state: StorageNotificationState,
   action: Action,
   plan: TriggeredActionPlan,
@@ -178,7 +156,7 @@ function applyPushTriggeredActionPlan(
   }
 }
 
-function applyPullTriggeredActionPlan(
+export function applyPullTriggeredActionPlan(
   state: StorageNotificationState,
   action: Action,
   isEffect: boolean,
@@ -209,7 +187,6 @@ interface CausalEdge {
 
 export interface StorageNotificationState {
   readonly triggerIndex: TriggerIndexState;
-  readonly getPullMode: () => boolean;
   readonly getDiagnosisEnabled: () => boolean;
   readonly getCollectTriggerTrace: () => boolean;
   readonly changeGroupToActionId: Map<ChangeGroup, string>;
@@ -228,149 +205,4 @@ export interface StorageNotificationState {
   readonly scheduleAffectedEffects: (
     action: Action,
   ) => TriggerTraceScheduledEffect[];
-}
-
-export function createStorageSubscription(
-  state: StorageNotificationState,
-): IStorageSubscription {
-  return {
-    next: (notification) => {
-      processStorageNotification(state, notification);
-      return { done: false };
-    },
-  };
-}
-
-export function processStorageNotification(
-  state: StorageNotificationState,
-  notification: StorageNotification,
-): void {
-  const space = notification.space;
-
-  if (!("changes" in notification)) {
-    return;
-  }
-
-  const sourceChangeGroup = notification.type === "commit"
-    ? notification.source?.changeGroup
-    : undefined;
-  const hasSourceChangeGroup = notification.type === "commit" &&
-    sourceChangeGroup !== undefined;
-  const pullMode = state.getPullMode();
-  const collectTriggerTrace = state.getCollectTriggerTrace();
-  const diagnosisEnabled = state.getDiagnosisEnabled();
-
-  let changeIndex = 0;
-  for (const change of notification.changes) {
-    changeIndex++;
-    state.recordCellUpdate(change);
-
-    if (
-      !hasRegisteredTriggers(state.triggerIndex)
-    ) {
-      continue;
-    }
-
-    const {
-      entity: spaceAndURI,
-      hasMatchingTriggerPaths,
-      triggeredActions,
-    } = collectTriggeredActionsForChange(
-      state.triggerIndex,
-      space,
-      change,
-    );
-
-    if (!hasMatchingTriggerPaths) {
-      continue;
-    }
-
-    const writerActionId = hasSourceChangeGroup &&
-        sourceChangeGroup !== undefined
-      ? state.changeGroupToActionId.get(sourceChangeGroup)
-      : undefined;
-    const triggerTraceEntry: TriggerTraceEntry | null = collectTriggerTrace
-      ? createTriggerTraceEntry({
-        notificationType: notification.type,
-        changeIndex,
-        matchedActionCount: triggeredActions.length,
-        pullMode,
-        writerActionId,
-        space,
-        change,
-      })
-      : null;
-
-    for (const action of triggeredActions) {
-      // Causal edge tracking for diagnosis.
-      if (
-        diagnosisEnabled && hasSourceChangeGroup &&
-        sourceChangeGroup !== undefined
-      ) {
-        const writerActionId = state.changeGroupToActionId.get(
-          sourceChangeGroup,
-        );
-        if (writerActionId) {
-          state.recordCausalEdge({
-            writer: writerActionId,
-            cell: spaceAndURI,
-            triggered: state.getActionId(action),
-            timestamp: performance.now(),
-          });
-        }
-      }
-
-      const actionChangeGroup = state.actionChangeGroups.get(action);
-      const actionId = state.getActionId(action);
-      const actionIsEffect = state.effects.has(action);
-      const actionType = actionIsEffect ? "effect" : "computation";
-      const pendingBefore = state.pending.has(action);
-      const dirtyBefore = state.dirty.has(action);
-      const isOwnCommitSource = notification.type === "commit" &&
-        notification.source !== undefined &&
-        state.inFlightSources.get(action)?.has(notification.source) === true;
-      const plan = planTriggeredNotificationAction({
-        pullMode,
-        isEffect: actionIsEffect,
-        dirtyBefore,
-        isOwnCommitSource,
-        hasSourceChangeGroup,
-        actionChangeGroup,
-        sourceChangeGroup,
-      });
-      let scheduledEffects: TriggerTraceScheduledEffect[] = [];
-
-      if (pullMode) {
-        scheduledEffects = applyPullTriggeredActionPlan(
-          state,
-          action,
-          actionIsEffect,
-          plan,
-        );
-      } else {
-        applyPushTriggeredActionPlan(state, action, plan);
-      }
-
-      triggerTraceEntry?.triggered.push(
-        createTriggerTraceActionRecord({
-          actionId,
-          actionType,
-          pullMode,
-          decision: plan.decision,
-          pendingBefore,
-          pendingAfter: state.pending.has(action),
-          dirtyBefore,
-          dirtyAfter: state.dirty.has(action),
-          scheduledEffects,
-        }),
-      );
-    }
-
-    if (
-      triggerTraceEntry &&
-      shouldRecordTriggerTraceEntry(triggerTraceEntry)
-    ) {
-      state.recordTriggerTrace(triggerTraceEntry);
-    }
-  }
 }
