+++ packages/runner/src/scheduler/subscriptions.ts
@@ -1,42 +1,25 @@
-import { getLogger } from "@commonfabric/utils/logger";
 import type { Cancel } from "../cancel.ts";
-import { toMemorySpaceAddress } from "../link-utils.ts";
 import type { IMemorySpaceAddress } from "../storage/interface.ts";
 import type { ChangeGroup } from "../storage/interface.ts";
 import { pendingDependencyCollectionMightAffect } from "./dependency-graph.ts";
-import {
-  type DependencyUpdateState,
-  setSchedulerDependencies,
-} from "./dependency-updates.ts";
+import { type DependencyUpdateState } from "./dependency-updates.ts";
 import {
   readsOverlapWrites,
   type WriterIndexState,
 } from "./scheduling-writes.ts";
-import {
-  replaceActionTriggerPaths,
-  setCancelForTriggerEntities,
-  type TriggerSubscriptionState,
-} from "./trigger-index.ts";
+import { type TriggerSubscriptionState } from "./trigger-index.ts";
 import { entityKey } from "./keys.ts";
 import type {
   Action,
-  PopulateDependencies,
   PopulateDependenciesEntry,
   ReactivityLog,
   SpaceScopeAndURI,
-  TelemetryAnnotations,
 } from "./types.ts";
-
-const logger = getLogger("scheduler", {
-  enabled: true,
-  level: "warn",
-});
 
 type SchedulerActionTypeState = {
   readonly isEffectAction: WeakMap<Action, boolean>;
   readonly effects: Set<Action>;
   readonly computations: Set<Action>;
-  readonly getPullMode: () => boolean;
   readonly getIdempotencyCheckMode: () => boolean;
   readonly queueExecution: () => void;
 };
@@ -98,7 +81,6 @@ export interface SchedulerSubscribeActionState {
   readonly dirty: ReadonlySet<Action>;
   readonly stale: ReadonlySet<Action>;
   readonly writeIndex: WriterIndexState;
-  readonly getPullMode: () => boolean;
   readonly setDebounce: (action: Action, ms: number) => void;
   readonly setNoDebounce: (action: Action, optOut: boolean) => void;
   readonly setThrottle: (action: Action, ms: number) => void;
@@ -123,227 +105,7 @@ export interface SchedulerSubscribeActionState {
   ) => void;
 }
 
-export function subscribeSchedulerAction(
-  state: SchedulerSubscribeActionState,
-  action: Action,
-  populateDependencies: PopulateDependencies | ReactivityLog,
-  options: SchedulerSubscribeOptions = {},
-): Cancel {
-  // Handle backwards-compatible ReactivityLog argument
-  let populateDependenciesEntry: PopulateDependenciesEntry;
-  let immediateLog: ReactivityLog | undefined;
-  if (typeof populateDependencies === "function") {
-    populateDependenciesEntry = populateDependencies;
-  } else {
-    // ReactivityLog provided directly - set up dependencies immediately
-    // (for backwards compatibility with code that passes reads/writes)
-    immediateLog = populateDependencies;
-    populateDependenciesEntry = immediateLog;
-  }
-  const {
-    isEffect = false,
-    debounce,
-    noDebounce,
-    throttle,
-  } = options;
-
-  updateSchedulerActionChangeGroup(
-    state.subscriptionState,
-    action,
-    options,
-  );
-
-  // Apply debounce settings if provided
-  if (debounce !== undefined) {
-    state.setDebounce(action, debounce);
-  }
-  if (noDebounce !== undefined) {
-    state.setNoDebounce(action, noDebounce);
-  }
-  // Apply throttle setting if provided
-  if (throttle !== undefined) {
-    state.setThrottle(action, throttle);
-  }
-
-  const actionIsEffect = updateSchedulerActionType(
-    state.subscriptionState,
-    action,
-    isEffect,
-    {
-      queueExecution: true,
-    },
-  );
-
-  // Track parent-child relationship if action is created during another action's execution
-  registerParentChildAction(state.subscriptionState, action);
-  const parent = state.actionParent.get(action);
-  if (
-    state.getPullMode() &&
-    !actionIsEffect &&
-    parent &&
-    state.activePullDemandActions.has(parent)
-  ) {
-    state.pullDemandedFirstRunComputations.add(action);
-    state.queueExecution();
-  }
-
-  logger.debug(
-    "schedule",
-    () => [
-      "Subscribing to action:",
-      action,
-      actionIsEffect ? "effect" : "computation",
-    ],
-  );
-
-  // Store the populateDependencies callback for use in execute()
-  state.populateDependenciesCallbacks.set(action, populateDependenciesEntry);
-
-  // In pull mode, newly subscribed computations can be the replacement for an
-  // already-running child graph (for example after a $TYPE change). Seed any
-  // statically declared writes immediately so existing effects can discover
-  // the new writer before the first execute() cycle.
-  if (
-    state.getPullMode() &&
-    !actionIsEffect &&
-    !immediateLog
-  ) {
-    const declaredWrites = (action as Partial<TelemetryAnnotations>).writes;
-    if (declaredWrites && declaredWrites.length > 0) {
-      setSchedulerDependencies(
-        state.dependencyUpdateState,
-        action,
-        {
-          reads: [],
-          shallowReads: [],
-          writes: declaredWrites.map(toMemorySpaceAddress),
-        },
-      );
-    }
-  }
-
-  // If a ReactivityLog was provided directly, set up dependencies immediately.
-  // This ensures writes are tracked right away for reverse dependency graph.
-  if (immediateLog) {
-    const { reads, shallowReads, log: schedulingLog } =
-      setSchedulerDependencies(
-        state.dependencyUpdateState,
-        action,
-        immediateLog,
-      );
-    state.updateDependents(action, schedulingLog);
-    const { entities } = replaceActionTriggerPaths(
-      state.triggerSubscriptionState,
-      action,
-      reads,
-      shallowReads,
-    );
-
-    // Register the cancel function for the latest trigger set.
-    setCancelForTriggerEntities(
-      state.triggerSubscriptionState,
-      action,
-      entities,
-    );
-  } else {
-    // Mark action for dependency collection before first run
-    state.pendingDependencyCollection.add(action);
-  }
-
-  // Mark as dirty and pending for first-time execution
-  // In pull mode this still doesn't mean execution: There needs to be an effect to trigger it.
-  state.markDirectDirty(action);
-  state.pending.add(action);
-  state.scheduledFirstTime.add(action);
-
-  if (
-    state.getPullMode() &&
-    !actionIsEffect &&
-    state.getSchedulingWrites(action)?.length
-  ) {
-    state.scheduleAffectedEffects(action);
-  }
-
-  // Emit telemetry for new subscription
-  const actionId = state.getActionId(action);
-  state.submitSubscribeTelemetry({
-    type: "scheduler.subscribe",
-    actionId,
-    isEffect: actionIsEffect,
-  });
-
-  return () => state.unsubscribe(action);
-}
-
-export function resubscribeSchedulerAction(
-  state: SchedulerSubscribeActionState,
-  action: Action,
-  log: ReactivityLog,
-  options: SchedulerResubscribeOptions = {},
-): void {
-  const { isEffect } = options;
-
-  updateSchedulerActionChangeGroup(
-    state.subscriptionState,
-    action,
-    options,
-  );
-
-  const { reads, shallowReads, log: schedulingLog } = setSchedulerDependencies(
-    state.dependencyUpdateState,
-    action,
-    log,
-  );
-
-  // Track action type for pull-based scheduling
-  // Once an action is marked as an effect, it stays an effect
-  const actionIsEffect = updateSchedulerActionType(
-    state.subscriptionState,
-    action,
-    isEffect,
-  );
-  const actionId = state.getActionId(action);
-
-  // Update reverse dependency graph after the action type is restored. In
-  // pull mode, registering a new edge to a live effect can be the moment a
-  // stale upstream computation becomes demanded.
-  if (state.getPullMode()) state.updateDependents(action, schedulingLog);
-
-  // Track parent-child relationship if action is created during another action's execution
-  // Only set if not already set (resubscribe can be called multiple times)
-  registerParentChildAction(state.subscriptionState, action, {
-    allowExisting: false,
-  });
-
-  const { entities, triggerPathsByEntity } = replaceActionTriggerPaths(
-    state.triggerSubscriptionState,
-    action,
-    reads,
-    shallowReads,
-  );
-
-  logger.debug("schedule-resubscribe", () => [
-    `Action: ${actionId}`,
-    `Entities: ${triggerPathsByEntity.size}`,
-    `Reads: ${reads.length}`,
-  ]);
-
-  setCancelForTriggerEntities(
-    state.triggerSubscriptionState,
-    action,
-    entities,
-  );
-
-  markEffectDirtyIfStaleInputs(
-    state,
-    action,
-    actionIsEffect,
-    reads,
-    shallowReads,
-  );
-}
-
-function markEffectDirtyIfStaleInputs(
+export function markEffectDirtyIfStaleInputs(
   state: SchedulerSubscribeActionState,
   action: Action,
   actionIsEffect: boolean,
@@ -355,7 +117,7 @@ function markEffectDirtyIfStaleInputs(
   // pull those computations and see fresh data.
   // Skip throttled computations - they'll trigger via storage changes when unthrottled.
   // Use isEffectAction instead of effects because unsubscribe() clears effects before run()
-  if (!state.getPullMode() || !actionIsEffect || state.stale.size === 0) {
+  if (!actionIsEffect || state.stale.size === 0) {
     return;
   }
 
@@ -421,7 +183,7 @@ export function updateSchedulerActionType(
   state: SchedulerActionTypeState,
   action: Action,
   isEffect: boolean | undefined,
-  options: { queueExecution?: boolean } = {},
+  options: { queueExecution?: boolean; queueComputation?: boolean } = {},
 ): boolean {
   if (isEffect) {
     state.isEffectAction.set(action, true);
@@ -440,7 +202,7 @@ export function updateSchedulerActionType(
     state.effects.delete(action);
     if (
       options.queueExecution &&
-      (!state.getPullMode() || state.getIdempotencyCheckMode())
+      (options.queueComputation ?? true)
     ) {
       state.queueExecution();
     }
